deps/v8/test/inspector/debugger/return-break-locations-expected.txt

Return break locations within function

Running test: testTailCall
[
    [0] : {
        columnNumber : 20
        lineNumber : 2
        scriptId : <scriptId>
        type : return
    }
    [1] : {
        columnNumber : 24
        lineNumber : 3
        scriptId : <scriptId>
        type : return
    }
    [2] : {
        columnNumber : 29
        lineNumber : 4
        scriptId : <scriptId>
        type : return
    }
    [3] : {
        columnNumber : 59
        lineNumber : 5
        scriptId : <scriptId>
        type : return
    }
    [4] : {
        columnNumber : 1
        lineNumber : 6
        scriptId : <scriptId>
        type : return
    }
]
